+++ DeepMuon/train/dist_train_vst.py
@@ -2,7 +2,7 @@
 Author: Airscker
 Date: 2022-07-19 13:01:17
 LastEditors: airscker
-LastEditTime: 2023-02-23 12:28:00
+LastEditTime: 2023-02-23 20:16:18
 Description: NULL
 
 Copyright (C) 2023 by Airscker(Yufeng), All Rights Reserved.
@@ -29,9 +29,8 @@ from torch.utils.data import DataLoader
 from torch.utils.data.distributed import DistributedSampler
 from torch.nn.parallel import DistributedDataParallel
 from torch.utils.tensorboard import SummaryWriter
-torch.set_default_tensor_type(torch.FloatTensor)
 torch.backends.cudnn.benchmark = True
-torch.manual_seed(3407)
+# torch.manual_seed(3407)
 
 try:
     from torch.distributed.fsdp import FullyShardedDataParallel as FSDP, FullStateDictConfig, StateDictType
@@ -258,7 +257,7 @@ def main(config_info, test_path=None):
                 tensorboard_plot(tr_eva_metrics, t+1, writer, 'train')
                 tensorboard_plot(ts_eva_metrics, t+1, writer, 'test')
             '''Save best model accoeding to the value of sota target'''
-            if ts_target != bestres:
+            if ts_target != bestres and not np.isnan(ts_target):
                 bestres = ts_target
                 if os.path.exists(best_checkpoint):
                     os.remove(best_checkpoint)
@@ -404,7 +403,7 @@ def train(device: Union[int, str, torch.device],
         - Gradient accumulation: Gradient accumulation steps
         - Mixed precision: Mixed precision training is allowed
         - Gradient resacle: Only available when mixed precision training is enabled, to avoid the gradient exploration/annihilation bring by fp16
-        - Gradient clip: Only available when mixed precision training is DISABLED
+        - Gradient clip: Using gradient value clip technique
     '''
     model.train()
     train_loss = 0
@@ -416,18 +415,18 @@ def train(device: Union[int, str, torch.device],
         x = x.to(device)
         y = y.reshape(-1).to(device)
         with autocast(enabled=fp16):
-            if (i+1) % gradient_accumulation != 0 and i+1 < batchs:
+            if (i+1) % gradient_accumulation != 0:
                 with model.no_sync():
                     pred = model(x)
                     loss = loss_fn(pred, y)
                     loss = loss/gradient_accumulation
                     grad_scalar.scale(loss).backward()
-            elif (i+1) % gradient_accumulation == 0 or i+1 == batchs:
+            elif (i+1) % gradient_accumulation == 0:
                 pred = model(x)
                 loss = loss_fn(pred, y)
                 loss = loss/gradient_accumulation
-                if grad_clip is not None and fp16 == False:
-                    torch.nn.utils.clip_grad_norm_(
+                if grad_clip is not None:
+                    torch.nn.utils.clip_grad_value_(
                         model.parameters(), grad_clip)
                 grad_scalar.scale(loss).backward()
                 grad_scalar.step(optimizer)
